+++ entities/Metrics.go
@@ -6,7 +6,7 @@ type Request struct {
 	ID string `gorm:"primary_key;type:uuid"`
 	Created time.Time
 	Method string
-	Status int
+	Status int64
 	Path string
 	Latency int64
 }
